+++ database/hash.go
@@ -611,32 +611,6 @@ func execHRandField(db *DB, args [][]byte) redis.Reply {
 }
 
 func execHScan(db *DB, args [][]byte) redis.Reply {
-	var count int = 10
-	var pattern string = "*"
-	if len(args) > 2 {
-		for i := 2; i < len(args); i++ {
-			arg := strings.ToLower(string(args[i]))
-			if arg == "count" {
-				if i+1 >= len(args) {
-					return &protocol.SyntaxErrReply{}
-				}
-				count0, err := strconv.Atoi(string(args[i+1]))
-				if err != nil {
-					return &protocol.SyntaxErrReply{}
-				}
-				count = count0
-				i++
-			} else if arg == "match" {
-				if i+1 >= len(args) {
-					return &protocol.SyntaxErrReply{}
-				}
-				pattern = string(args[i+1])
-				i++
-			} else {
-				return &protocol.SyntaxErrReply{}
-			}
-		}
-	}
 	if len(args) < 2 {
 		return &protocol.SyntaxErrReply{}
 	}
@@ -647,6 +621,10 @@ func execHScan(db *DB, args [][]byte) redis.Reply {
 	}
 	if dict == nil {
 		return emptyScanReply()
+	}
+	count, pattern, optErr := parseTypeScanOptions(args, 2)
+	if optErr != nil {
+		return optErr
 	}
 	cursor, err := strconv.Atoi(string(args[1]))
 	if err != nil {